# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Lifio API
  description: |
    Backend API for the Lifio personal time-tracking iOS app.

    ## Auth flow (passwordless email-OTP)
    1. **Request OTP** — `POST /api/v1/auth/otp/request` with an email address. Always returns 202
       (prevents user enumeration). The server upserts the user, generates a 6-digit code,
       and sends it via email (console in dev/test, AWS SES in production).
    2. **Verify OTP** — `POST /api/v1/auth/otp/verify` with the email and 6-digit code. On success,
       returns a JWT access token (15 min TTL) and an opaque refresh token.
    3. **Refresh** — `POST /api/v1/auth/refresh` with the refresh token and `X-Device-Id`.
       Issues a new token pair within the sending device's refresh family and revokes the
       old token. Reuse of a revoked token revokes only that device's family — other
       devices' sessions keep working.
    4. **Logout** — `POST /api/v1/auth/logout` with a Bearer access token and `X-Device-Id`.
       Revokes only the calling device's refresh family; other devices' sessions remain valid.
    5. **Me** — `GET /api/v1/auth/me` with a Bearer access token. Returns the user profile.

    ## Auth flow (Sign in with Apple)
    - **Apple** — `POST /api/v1/auth/apple` with Apple's identity-token JWT (RS256). The server
      verifies it against Apple's JWKS (iss/aud/exp) and upserts a user keyed by Apple's
...
       The route is always registered; without `APPLE_CLIENT_ID` it returns 503
...

...

...

...

...
servers:
...
paths:
...
  /api/v1/auth/otp/request:
    post:
      operationId: requestOTP
      summary: Request an OTP code
      description: |
        Sends a 6-digit OTP code to the given email address. The user account is
        created automatically if it does not exist. Always returns HTTP 202 to
        prevent user enumeration — a 202 response does NOT guarantee that an email
        was sent (internal errors are silently swallowed).
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
            example:
              email: user@example.com
      responses:
        "202":
          description: Request accepted (email may or may not have been sent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: invalid_body
                  message: Invalid request body
        "429":
          description: Rate limited (3 req/min per IP+email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
  /api/v1/auth/otp/verify:
    post:
      operationId: verifyOTP
      summary: Verify an OTP code
      description: |
        Verifies the 6-digit OTP code for the given email. On success, marks the
        user as email-verified and returns a JWT access token + opaque refresh token.
        The refresh token is recorded against the `X-Device-Id` device's refresh family.
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
            example:
              email: user@example.com
              code: "123456"
      responses:
        "200":
          description: OTP verified successfully — tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body (invalid email or code format) or missing device header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_body:
                  summary: Invalid email or code format
                  value:
                    error:
                      code: invalid_body
                      message: Invalid email or code
                invalid_request:
                  summary: Missing X-Device-Id header
                  value:
                    error:
                      code: invalid_request
                      message: X-Device-Id header is required
        "401":
          description: Invalid, expired, or exhausted OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_otp:
                  summary: Invalid code
                  value:
                    error:
                      code: invalid_otp
                      message: Invalid or expired code
                otp_expired:
                  summary: Code expired
                  value:
                    error:
                      code: otp_expired
                      message: Code has expired. Request a new one.
                otp_attempts_exceeded:
                  summary: Too many failed attempts
                  value:
                    error:
                      code: otp_attempts_exceeded
                      message: Too many incorrect attempts. Request a new code.
        "429":
          description: Rate limited (5 req/min per IP+email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many attempts. Please try again later.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/apple:
    post:
      operationId: signInWithApple
      summary: Sign in / register with Apple
      description: |
        Verifies Apple's identity-token JWT (RS256) against Apple's JWKS, validating the
        issuer, audience (the app's Bundle ID), and expiry. Upserts a user keyed by Apple's
        stable `sub` claim (Apple may withhold the email after the first authorization) and
        returns the same access + refresh token pair as the OTP flow. Apple users are
        considered email-verified. The refresh token is recorded against the
        `X-Device-Id` device's refresh family.

        This endpoint is always registered. When the server is not configured
        with `APPLE_CLIENT_ID`, it returns HTTP 503 (`apple_not_configured`).
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleSignInRequest"
            example:
              identity_token: eyJraWQiOi...id-token...
      responses:
        "200":
          description: Apple sign-in successful — tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body or missing device header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_body:
                  summary: Identity token missing or malformed
                  value:
                    error:
                      code: invalid_body
                      message: Identity token is required
                invalid_request:
                  summary: Missing X-Device-Id header
                  value:
                    error:
                      code: invalid_request
                      message: X-Device-Id header is required
        "401":
          description: Apple identity token failed verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: invalid_apple_token
                  message: Invalid Apple identity token
        "429":
          description: Rate limited (per IP)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: Too many requests. Please try again later.
        "503":
          description: Sign in with Apple is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: apple_not_configured
                  message: Sign in with Apple is not configured
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access + refresh token pair.
        The old refresh token is revoked and the new one stays within the
        `X-Device-Id` device's refresh family. Rotation never affects other
        devices' sessions. There is deliberately NO device limit: no 409
        `device_limit` response exists (device quota counting is deferred to
        issue #46 — do not file its absence as a bug).
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: abc123...
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body or missing device header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_body:
                  summary: Refresh token missing or malformed
                  value:
                    error:
                      code: invalid_body
                      message: Refresh token is required
                invalid_request:
                  summary: Missing X-Device-Id header
                  value:
                    error:
                      code: invalid_request
                      message: X-Device-Id header is required
        "401":
          description: |
            Invalid, revoked (reuse), or TTL-expired refresh token. Reuse detection is
            scoped per-device: only the offending device's refresh family is revoked —
            other devices' sessions keep working. After a revoked or TTL-expired family,
            the device must re-authenticate (the client locks to the auth gate with all
            local data intact; nothing is wiped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_refresh:
                  summary: Token not found or presented from the wrong device
                  value:
                    error:
                      code: invalid_refresh
                      message: Invalid refresh token
                token_reuse:
                  summary: Token reuse detected (offending device's family revoked)
                  value:
                    error:
                      code: token_reuse
                      message: Token has been revoked. The device's sessions have been invalidated.
                refresh_expired:
                  summary: Refresh family TTL exhausted (re-authenticate, local data intact)
                  value:
                    error:
                      code: refresh_expired
                      message: Refresh token has expired. Please sign in again.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Logout
      description: |
        Revokes only the refresh family identified by the request's `X-Device-Id`;
        other devices' sessions remain valid. Requires a valid Bearer access token.
        After logout the client keeps all local data intact (logout never wipes).
      tags:
        - Auth
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      responses:
        "204":
          description: Logout successful — the calling device's refresh family revoked
        "400":
          description: Missing X-Device-Id header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: invalid_request
                  message: X-Device-Id header is required
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: unauthorized
                  message: Missing authorization header
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current user
      description: |
        Returns the authenticated user's profile.
        Requires a valid Bearer access token.
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: unauthorized
                  message: Invalid or expired access token
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Categories ----------
  /api/v1/categories:
    get:
      operationId: listCategories
      summary: List categories
      description: |
        Returns the authenticated user's categories, ordered by name.
        This is a full pull: the handler ignores `modified_since` (delta pulls
        are entries-only). Adding `modified_since` support here is a backend
        follow-up, not part of this contract.
      tags:
        - Categories
      security:
        - BearerAuth: []
      responses:
        "200":
          description: The user's categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createCategory
      summary: Create a category
      description: |
        Creates a category. `id` is a client-generated UUID v7 (idempotent on
        replay). A case-insensitive name collision returns 409 `category_exists`.
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "201":
          description: Category created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "200":
          description: Idempotent replay — a category with this `id` already exists and is returned unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/InvalidBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/categories/{id}:
    get:
      operationId: getCategory
      summary: Get a category
      description: Returns a single category by id.
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateCategory
      summary: Update a category
      description: |
        Updates a category (last-write-wins on `updated_at`). A name change that
        collides with another category returns 409 `category_exists`.
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: The updated category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/InvalidBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteCategory
      summary: Delete a category
      description: |
        Hard-deletes a category. The `entry_categories` join rows cascade (the
        tag disappears from every entry); entries themselves are unaffected —
        their text, notes, and timings stay intact.
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Category deleted (tag-join rows cascaded).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Entries ----------
  /api/v1/entries:
    get:
      operationId: listEntries
      summary: List entries
      description: |
        Returns the authenticated user's entries, newest first, with filters and
        cursor pagination. Each entry carries its own `activity_text`, ordered
        category tags, and notes — no record is resolved at query time. Used by
        History; entries sync cross-device from the start.
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound on `started_at`.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive upper bound on `started_at`.
        - name: category_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to entries tagged with this category (via `entry_categories`).
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor returned by a previous request.
        - name: modified_since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Delta pull-sync: return only entries with `updated_at` strictly
            greater than this RFC 3339 timestamp. Absent/empty = full pull.
      responses:
        "200":
          description: A page of entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createEntry
      summary: Create a time entry
      description: |
        Creates a time entry. `activity_text` is required (trimmed, non-empty,
        max 60 chars; case-sensitive identity); `category_ids` are stored on
        the entry via the `entry_categories` join (order preserved, unknown or
        non-owned ids rejected with 422). `ended_at` null/omitted = running.
        `id` is client-generated UUID v7 (idempotent on replay).
      tags:
        - Entries
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryCreate"
      responses:
        "201":
          description: Entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "200":
          description: Idempotent replay — an entry with this `id` already exists and is returned unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          $ref: "#/components/responses/InvalidBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/entries/recents:
    get:
      operationId: listRecents
      summary: List recent entry texts
      description: |
        Returns up to `limit` entries (default 6, max 20) — the newest entry
        per exact `activity_text` (`GROUP BY activity_text`, newest
        `started_at` wins per group, `id DESC` tiebreak), newest-first. Each
        item is a full entry with its own ordered category tags, so a fresh
        device can render Recents (exact text + first-position category icon)
        before its first full entry pull. Identity is byte-exact trimmed text,
        case-sensitive (`Gym` ≠ `GYM`).
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
          description: Max exact-text groups to return.
      responses:
        "200":
          description: The recents entries, newest group first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/deletions:
    get:
      operationId: listDeletions
      summary: List deletion tombstones
      description: |
        Returns the authenticated user's deletion tombstones ordered by
        `deleted_at` ascending (cursor-friendly). Every hard DELETE of a
        category or entry records one tombstone, so a device that missed the
        delete converges by removing its local copy (an entry tombstone also
        removes its `entry_categories` joins; a category tombstone removes its
        joins and leaves entries intact). Recreating an id clears its
        tombstone, so a tombstone older than a live record's `updated_at` is
        stale and ignored. Tombstones are retained (no GC); `deleted_at`
        enables a future retention policy.
      tags:
        - Sync
      security:
        - BearerAuth: []
      parameters:
        - name: deleted_since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Return only tombstones with `deleted_at` strictly greater than
            this RFC 3339 timestamp. Absent/empty = full list.
      responses:
        "200":
          description: The user's deletion tombstones, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deletion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/otp/verify`, `/api/v1/auth/apple`, or `/api/v1/auth/refresh`.
        Format: `Authorization: Bearer <token>`
...
  schemas:
...
    OtpRequest:
      type: object
      description: Request body for requesting an OTP code.
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: User's email address
          example: user@example.com
...